Temporarily removed use of "L" for labels since it breaks assem/disassem
Disassembly should create jump labels in the source file. The previous
version simply added an "L" before the address of each line. However,
this was not compatible with repeated assembly and disassembly of the
same code at this time. For that reason, the "L" was removed from the
labels leaving them as plain hex labels. This will be fixed in the future
when labels are handled properly.