Added XOR Drawing; Removed hard-coded default; Improved Help; Time Control
The addition of XOR drawing was demonstrated by the new XOR Box example.
The default program is no longer hard-coded into the text area. It is now
loaded from one of the examples which can be easily changed.
The Tetris example was added, but it does not work yet.
A number of time control elements have been added in anticipation of an
actual clock. There is a new "Clock Time" (CT) variable to represent the
duration of the callback delay. Similarly a "Steps per clock Tick" (ST)
variable was added to represent the number of program steps executed in
a single clock tick. At this time, ST is 1000 which represents the number
of steps to be executed for each "Run" click (to be consistent with
earlier versions).